+++ src/main/java/net/silentchaos512/scalinghealth/network/SHMechanicsPacket.java
@@ -2,7 +2,7 @@ package net.silentchaos512.scalinghealth.network;
 
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.chat.TextComponent;
-import net.minecraftforge.fmllegacy.network.NetworkEvent;
+import net.minecraftforge.network.NetworkEvent;
 import net.silentchaos512.scalinghealth.ScalingHealth;
 import net.silentchaos512.scalinghealth.resources.mechanics.*;