+++ src/main/java/net/silentchaos512/scalinghealth/world/SHWorldFeatures.java
@@ -15,6 +15,7 @@ import net.minecraft.world.level.levelgen.placement.CountPlacement;
 import net.minecraft.world.level.levelgen.placement.HeightRangePlacement;
 import net.minecraft.world.level.levelgen.placement.InSquarePlacement;
 import net.minecraft.world.level.levelgen.placement.PlacedFeature;
+import net.minecraft.world.level.levelgen.structure.templatesystem.RuleTest;
 import net.minecraftforge.event.world.BiomeLoadingEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -30,8 +31,14 @@ import java.util.function.Supplier;
 @Mod.EventBusSubscriber(modid = ScalingHealth.MOD_ID)
 public class SHWorldFeatures {
     private static final List<OreSpawnInfo> ORES = ImmutableList.of(
-            new OreSpawnInfo("heart_crystal_ore", Registration.HEART_CRYSTAL_ORE, 6, 1, 28, EnabledFeatures::hpCrystalsOreGenEnabled),
-            new OreSpawnInfo("power_crystal_ore", Registration.POWER_CRYSTAL_ORE, 5, 1, 28, EnabledFeatures::powerCrystalsOreGenEnabled)
+            new OreSpawnInfo("heart_crystal_ore", Registration.HEART_CRYSTAL_ORE, 4, 1,
+                    OreFeatures.STONE_ORE_REPLACEABLES, 100, EnabledFeatures::hpCrystalsOreGenEnabled),
+            new OreSpawnInfo("power_crystal_ore", Registration.POWER_CRYSTAL_ORE, 3, 1,
+                    OreFeatures.STONE_ORE_REPLACEABLES, 100, EnabledFeatures::powerCrystalsOreGenEnabled),
+            new OreSpawnInfo("deepslate_heart_crystal_ore", Registration.DEEPLSATE_HEART_CRYSTAL_ORE, 6, 1,
+                    OreFeatures.DEEPSLATE_ORE_REPLACEABLES, 64, EnabledFeatures::hpCrystalsOreGenEnabled),
+            new OreSpawnInfo("deepslate_power_crystal_ore", Registration.DEEPSLATE_POWER_CRYSTAL_ORE, 5, 1,
+                    OreFeatures.DEEPSLATE_ORE_REPLACEABLES, 64, EnabledFeatures::powerCrystalsOreGenEnabled)
     );
 
     @SubscribeEvent(priority = EventPriority.HIGH)
@@ -46,20 +53,20 @@ public class SHWorldFeatures {
         private final Supplier<PlacedFeature> feature;
         private final BooleanSupplier test;
 
-        public OreSpawnInfo(String name, Supplier<Block> block, int size, int count, int height, BooleanSupplier test) {
+        public OreSpawnInfo(String name, Supplier<Block> block, int size, int count, RuleTest replacing, int height, BooleanSupplier test) {
             this.feature = Suppliers.memoize(() -> {
                 ConfiguredFeature<? ,? > f = Registry.register(
                     BuiltinRegistries.CONFIGURED_FEATURE,
                     ScalingHealth.getId(name),
                     Feature.ORE.configured(new OreConfiguration(
-                            OreFeatures.STONE_ORE_REPLACEABLES,
+                            replacing,
                             block.get().defaultBlockState(),
                             size
                     )));
                 return Registry.register(BuiltinRegistries.PLACED_FEATURE,
                         ScalingHealth.getId(name),
                         f.placed(
-                                HeightRangePlacement.uniform(VerticalAnchor.absolute(10), VerticalAnchor.absolute(height)),
+                                HeightRangePlacement.uniform(VerticalAnchor.bottom(), VerticalAnchor.absolute(height)),
                                 InSquarePlacement.spread(),
                                 CountPlacement.of(count)));
             });